+++ src/loader.cpp
@@ -7,6 +7,7 @@
 void InitializeUnifiedAttack();
 void RangedAutoStop(Player*);
 void RegisterAdaptiveAttackHooks();
+void RegisterAdaptiveAttackOptions();
 
 class adaptive_autoattack_player : public PlayerScript
 {
@@ -31,4 +32,5 @@ void Addmod_adaptive_autoattackScripts()
     InitializeUnifiedAttack();
     new adaptive_autoattack_player();
     RegisterAdaptiveAttackHooks();
+    RegisterAdaptiveAttackOptions();
 }